Addressed feed back from Pat and Marty, noticed and fixed assumption in generate_changelog

+++ scripts/release.sh
@@ -2,7 +2,6 @@
 
 RELEASE_VERSION=$1
 RELEASE_DATE=$2
-cd ..
 
 #Get linode/manager repo name from remote branches
 GET_ORIGIN=$(git remote -v | grep git@github.com:linode/manager.git)
@@ -16,7 +15,7 @@ git rebase $ORIGIN/develop
 git checkout -b release-$RELEASE_VERSION
 
 #Generate changelog
-python generate_changelog.py $RELEASE_VERSION $RELEASE_DATE
+python generate_changelog.py $RELEASE_VERSION $RELEASE_DATE $ORIGIN
 
 #Create release tag & bump package.json version
 yarn version --new-version $RELEASE_VERSION
@@ -24,8 +23,18 @@ yarn version --new-version $RELEASE_VERSION
 #Commit updated version and new changelog
 git commit -m "Cloud Manager version $RELEASE_VERSION - $RELEASE_DATE"
 
-#Git delete previous release branch
+#Get previous release branch
 PREVIOUS_RELEASE=$(git ls-remote --heads $ORIGIN | egrep -o 'release-\d*\.\d*\.\d*')
+
+#Pause to manually check and edit change log, updated release version
+read -p "WARNING: The upstream release-$PREVIOUS_RELEASE branch will be deleted and replaced with local branch release-$RELEASE_VERSION. Are you ready to release? (y/n) " -n 1 -r
+
+if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+    echo "You will need to delete release-$PREVIOUS_RELEASE manually and push release-$RELEASE_VERSION "
+    exit 0
+fi
+
+#Git delete previous release branch
 git push $ORIGIN  --delete $PREVIOUS_RELEASE
 
 #Push release
